+++ H_euclidean.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 extern default_random_engine generator;
 
-H_euclidean::H_euclidean(int w, int dimension, double mean, double variance):H(dimension, mean, variance) {
+H_euclidean::H_euclidean(int w, double dimension, double mean, double variance):H(dimension, mean, variance) {
   this->t = generateT(0.0, 400.0);
   this->w = 400;
 }
